backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
distlib==0.3.8
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==4.2.14
django-allauth==64.2.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-sendgrid-v5==1.2.3
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.7
env==0.1.0
filelock==3.15.4
gunicorn==23.0.0
idna==3.9
inflection==0.5.1
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
sendgrid==6.11.0
sendgrid-django==4.2.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.3
whitenoise==6.7.0
